refactor(merge-strategies): extract strategies into core/merge-strategies and add LWW strategy
Moves MergeStrategy interface, MergeMeta, LamportMeta, and
createLamportStrategy out of useSharedState.ts into a dedicated
core/merge-strategies/ folder. Adds a new createLastWriteWinsStrategy
(wall-clock with peer ID tiebreaker). All types and factories remain
publicly exported from the package root.
