+++ packages/react-p2p/src/core/merge-strategies/index.ts
@@ -1,0 +1,3 @@
+export { createLamportStrategy, type LamportMeta } from './lamport';
+export { createLastWriteWinsStrategy, type LastWriteWinsMeta } from './lastWriteWins';
+export type { MergeMeta, MergeStrategy } from './types';
