+++ packages/react-p2p/src/hooks/useSharedState.ts
@@ -1,31 +1,12 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useRoom } from '..';
 import type { JSONSerializable } from '../context/Room';
-
-export type MergeMeta = Record<string, JSONSerializable>;
-
-export interface MergeStrategy<
-  TState extends JSONSerializable = JSONSerializable,
-  TMeta extends MergeMeta = MergeMeta,
-> {
-  /** Initial metadata before any sync or update. */
-  initialMeta: TMeta;
-
-  /** Produce meta when this peer sets state (e.g. timestamp, version). */
-  createMeta(): TMeta;
-
-  /**
-   * Decide how to merge incoming state with current.
-   * Return new { state, meta } to apply, or null to keep current.
-   */
-  merge(
-    currentState: TState | null,
-    currentMeta: TMeta,
-    incomingState: TState | null,
-    incomingMeta: TMeta,
-    senderId: string
-  ): { state: TState | null; meta: TMeta } | null;
-}
+import {
+  createLamportStrategy,
+  type LamportMeta,
+  type MergeMeta,
+  type MergeStrategy,
+} from '../core/merge-strategies';
 
 type SharedStatePayload<TState extends JSONSerializable, TMeta extends MergeMeta = MergeMeta> = {
   key: string;
@@ -44,65 +25,6 @@ function isStateRequest(
   return (
     typeof data === 'object' && data !== null && 'type' in data && data.type === 'state-request'
   );
-}
-
-/**
- * Metadata carried by the Lamport clock strategy.
- * `clock` is a monotonically increasing logical counter; `tiebreaker` is the
- * writing peer's ID, used to deterministically resolve equal-clock conflicts.
- */
-export type LamportMeta = { clock: number; tiebreaker: string };
-
-/**
- * Creates a Lamport logical-clock merge strategy.
- *
- * Unlike wall-clock timestamps, Lamport clocks do not rely on peers having
- * synchronised system clocks. The clock advances monotonically: it increments
- * on every local write and is fast-forwarded to `max(local, incoming)` on
- * every receive, so causal ordering is preserved across all peers.
- *
- * Concurrent writes (same clock value) are broken deterministically by
- * comparing the writing peer's ID as a string, so every peer reaches the same
- * result independently.
- *
- * @param getPeerId - A function that returns the local peer's ID at call time.
- *   Pass a ref-backed getter so the strategy stays stable even before the
- *   signalling handshake completes.
- */
-export function createLamportStrategy(
-  getPeerId: () => string
-): MergeStrategy<JSONSerializable, LamportMeta> {
-  let localClock = 0;
-
-  return {
-    initialMeta: { clock: 0, tiebreaker: getPeerId() },
-
-    createMeta(): LamportMeta {
-      return { clock: ++localClock, tiebreaker: getPeerId() };
-    },
-
-    merge(
-      _currentState,
-      currentMeta,
-      incomingState,
-      incomingMeta,
-      senderId
-    ): { state: JSONSerializable | null; meta: LamportMeta } | null {
-      localClock = Math.max(localClock, incomingMeta.clock);
-
-      const clockWins = incomingMeta.clock > currentMeta.clock;
-      const incomingTiebreaker = incomingMeta.tiebreaker || senderId;
-      const currentTiebreaker = currentMeta.tiebreaker || getPeerId();
-
-      const tiebreakWins =
-        incomingMeta.clock === currentMeta.clock &&
-        incomingTiebreaker > currentTiebreaker;
-
-      return clockWins || tiebreakWins
-        ? { state: incomingState, meta: { ...incomingMeta, tiebreaker: incomingTiebreaker } }
-        : null;
-    },
-  };
 }
 
 /**
