+++ packages/react-p2p/src/index.ts
@@ -1,10 +1,12 @@
 // Export Room context and hook
 export { Room, RoomContext, type RoomContextValue } from './context/Room';
-export { useRoom } from './hooks/useRoom';
 export {
   createLamportStrategy,
+  createLastWriteWinsStrategy,
   type LamportMeta,
+  type LastWriteWinsMeta,
   type MergeMeta,
   type MergeStrategy,
-  useSharedState,
-} from './hooks/useSharedState';
+} from './core/merge-strategies';
+export { useRoom } from './hooks/useRoom';
+export { useSharedState } from './hooks/useSharedState';
